+++ src/pathogen/progression_categories/pc_symptomatic.jl
@@ -39,13 +39,13 @@ end
 function calculate_progression(individual::Individual, tick::Int16, dp::Symptomatic, rng::Xoshiro)
 
     # Calculate the time to infectiousness
-    infectiousness_onset = tick + Int16(1) + rand_val(dp.exposure_to_infectiousness_onset, rng)
+    infectiousness_onset = Int16(tick + 1 + rand_val(dp.exposure_to_infectiousness_onset, rng))
 
     # Calculate the time to symptom onset
-    symptom_onset = infectiousness_onset + rand_val(dp.infectiousness_onset_to_symptom_onset, rng)
+    symptom_onset = Int16(infectiousness_onset + rand_val(dp.infectiousness_onset_to_symptom_onset, rng))
 
     # Calculate the time to recovery
-    recovery = symptom_onset + rand_val(dp.symptom_onset_to_recovery, rng)
+    recovery = Int16(symptom_onset + rand_val(dp.symptom_onset_to_recovery, rng))
 
     return DiseaseProgression(
         exposure = tick,
